requirements.txt
requests==2.26.0
pytest==7.1.2
python-dotenv==0.19.1